[Refactor] NoteMain, NoteAside로 컴포넌트 분리
